pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iceberg-engine-benchmarking"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "boto3>=1.34",
]

[project.optional-dependencies]
duckdb = [
    "duckdb>=1.1",
    "pyarrow>=14.0",
]
spark = [
    "pyspark>=3.5,<4.0",
    "iceberg-engine-benchmarking[duckdb]",
]
local = [
    # Only needed when using the local catalog (writes via PyIceberg instead of DuckDB)
    "pyiceberg[s3,pyarrow]>=0.7",
    "pyarrow>=14.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]
all = [
    "iceberg-engine-benchmarking[duckdb,spark,local]",
]

[tool.hatch.build.targets.wheel]
packages = ["catalogs", "engines", "benchmarks", "setup"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
